+++ gg-agenda-api/src/main/java/gg/agenda/api/admin/agendateam/service/AgendaTeamAdminService.java
@@ -97,6 +97,6 @@ public class AgendaTeamAdminService {
 		List<AgendaTeamProfile> agendaTeamProfiles = agendaTeamProfileAdminRepository
 			.findAllByAgendaTeamAndIsExistIsTrue(agendaTeam);
 		agendaTeamProfiles.forEach(AgendaTeamProfile::changeExistFalse);
-		agendaTeam.cancelTeam(agendaTeam.getStatus());
+		agendaTeam.adminCancelTeam();
 	}
 }
