+++ gg-agenda-api/src/main/java/gg/agenda/api/user/agendateam/service/AgendaTeamService.java
@@ -15,7 +15,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import gg.agenda.api.user.SnsMessageUtil;
 import gg.agenda.api.user.agendateam.controller.request.TeamCreateReqDto;
 import gg.agenda.api.user.agendateam.controller.request.TeamKeyReqDto;
 import gg.agenda.api.user.agendateam.controller.request.TeamUpdateReqDto;
@@ -23,6 +22,7 @@ import gg.agenda.api.user.agendateam.controller.response.MyTeamSimpleResDto;
 import gg.agenda.api.user.agendateam.controller.response.TeamDetailsResDto;
 import gg.agenda.api.user.agendateam.controller.response.TeamKeyResDto;
 import gg.agenda.api.user.ticket.service.TicketService;
+import gg.agenda.api.utils.SnsMessageUtil;
 import gg.auth.UserDto;
 import gg.data.agenda.Agenda;
 import gg.data.agenda.AgendaProfile;
@@ -295,48 +295,5 @@ public class AgendaTeamService {
 		agendaTeamRepository.save(agendaTeam);
 	}
 
-	public void slackConfirmAgendaTeam(Agenda agenda, AgendaTeam newTeam) {
-		List<AgendaTeamProfile> agendaTeamProfiles = agendaTeamProfileRepository.findByAgendaTeamAndIsExistTrue(
-			newTeam);
-		String message = snsMessageUtil.confirmTeamMessage(agenda, newTeam);
-		agendaTeamProfiles.stream().map(atp -> atp.getProfile().getIntraId())
-			.forEach(intraId -> messageSender.send(intraId, message));
-		String toHostMessage = snsMessageUtil.agendaHostMessage(agenda);
-		if (agenda.getHostIntraId() != null) {
-			messageSender.send(agenda.getHostIntraId(), toHostMessage);
-		}
-	}
-
-	public void slackCancelAgendaTeam(Agenda agenda, AgendaTeam newTeam) {
-		List<AgendaTeamProfile> agendaTeamProfiles = agendaTeamProfileRepository.findByAgendaTeamAndIsExistTrue(
-			newTeam);
-		String message = snsMessageUtil.cancelTeamMessage(agenda, newTeam);
-		agendaTeamProfiles.stream().map(atp -> atp.getProfile().getIntraId())
-			.forEach(intraId -> messageSender.send(intraId, message));
-	}
-
-	public void slackCancelByAgendaConfirm(Agenda agenda, List<AgendaTeam> failTeam) {
-		List<AgendaTeamProfile> agendaTeamProfiles = agendaTeamProfileRepository.findByAgendaTeamInAndIsExistTrue(
-			failTeam);
-		String message = snsMessageUtil.failTeamMessage(agenda);
-		agendaTeamProfiles.stream().map(atp -> atp.getProfile().getIntraId())
-			.forEach(intraId -> messageSender.send(intraId, message));
-	}
-
-	public void slackAttendTeamMate(Agenda agenda, AgendaTeam agendaTeam, String userIntraId) {
-		List<AgendaTeamProfile> agendaTeamProfiles = agendaTeamProfileRepository.findByAgendaTeamAndIsExistTrue(
-			agendaTeam);
-		String message = snsMessageUtil.attendTeamMateMessage(agenda, agendaTeam, userIntraId);
-		agendaTeamProfiles.stream().map(atp -> atp.getProfile().getIntraId())
-			.forEach(intraId -> messageSender.send(intraId, message));
-	}
-
-	public void slackLeaveTeamMate(Agenda agenda, AgendaTeam agendaTeam, String userIntraId) {
-		List<AgendaTeamProfile> agendaTeamProfiles = agendaTeamProfileRepository.findByAgendaTeamAndIsExistTrue(
-			agendaTeam);
-		String message = snsMessageUtil.leaveTeamMateMessage(agenda, agendaTeam, userIntraId);
-		agendaTeamProfiles.stream().map(atp -> atp.getProfile().getIntraId())
-			.forEach(intraId -> messageSender.send(intraId, message));
-	}
 }
