+++ gg-agenda-api/src/main/java/gg/agenda/api/utils/SnsMessageUtil.java
@@ -1,4 +1,4 @@
-package gg.agenda.api.user;
+package gg.agenda.api.utils;
 
 import org.springframework.stereotype.Component;
